avatar fixes

+++ index.php
@@ -8,7 +8,6 @@
     <link rel="stylesheet" href="css/responsive.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css">
-
 </head>
 <body>
 
@@ -32,14 +31,12 @@
                     <li><a href="#">Gallery</a></li>
                     <li><a href="#">Contact</a></li>
                 </ul>
-
-                <?php
                 
+                <?php
                 require_once 'config.php';
 
                 if (isset($_SESSION['customer_email'])) {
 
-                    // Prepare and execute the query to fetch the user's picture path
                     $customer_email = $_SESSION['customer_email'];
                     $sql = "SELECT picture FROM customers WHERE email = '$customer_email'";
                     $result = mysqli_query($conn, $sql);
@@ -47,17 +44,23 @@
                     if ($result && mysqli_num_rows($result) > 0) {
                         $row = mysqli_fetch_assoc($result);
                         $user_picture_path = $row['picture'];
-
-                        echo '<div class="user-picture">';
-                        echo '<a href="logout.php"> <img src="' . $user_picture_path . '" alt="User Picture" class="avatar">';
-                        echo '</div>';
+                
+                        // Check if the picture path is NULL or the image doesn't load
+                        if ($user_picture_path && file_exists($user_picture_path)) {
+                            echo '<div class="user-picture">';
+                            echo '<a href="logout.php"> <img src="' . $user_picture_path . '" alt="User Picture" class="avatar"></a>';
+                            echo '</div>';
+                        } else {
+                            echo '<div class="user-picture">';
+                            echo '<a href="logout.php"><img src="Images/users/avatar_placeholder.png" alt="User Picture" class="avatar"></a>';
+                            echo '</div>';
+                        }
                     } else {
                         echo 'User picture not found.';
                     }
-
                     mysqli_close($conn);
                 } else {
-                    // If user is not signed in, display sign in button
+                    // if user not signed in, then:
                     echo '<div class="sign-in-up-btn">';
                     echo '<a href="login.php" class="custom-btn">Sign In Now</a>';
                     echo '</div>';
@@ -66,7 +69,6 @@
                 <div class="btn">
                     <i class="fas fa-bars menu-btn"></i>
                 </div>
-
             </nav>
         </div>
     </header>
